Bubbles for pending files
file_message_view_in_pending.xml and file_message_view_out_pending.xml are
complete and can are used for files that are still transferring

The SlowMessageLoader now generates file messages too.

There's a new interface: MonkeyFile. All FileHolders: MonkeyFileHolder,
MonkeyImageHolder and MonkeyAudioHolder must implement it so that when they are
bound to a file message they can be resent if necesary.