Fix issue when playing multiple audio files
There was a bug that would leave an audio unplayable if you interrupted it to
play a different one. To fix this, it is necessary to rebind the interrupted
audio so that it can show the play button. startAudioHolderPlayer now rebinds
both the new playing item and the old one.

removed excess logging in DefaultVoiceNotePlayer.