small changes in InputView and MonkeyAdapter
inputListener attribute is now declared in BaseInputView instead of
TextInputView. in addition to having a leftButtonView attribute, a
rightButtonView attribute was included, although right now it is not used.

MonkeyAdapter's constructor now only receives one argument, the context.
It initializes the ArrayList with an empty list. With this change, the only way
to add messages to the RecyclerView is with the MonkeyAdapter API