Minor changes in DefaultVoiceNotePlayer
the releasePlayer method now calls the onPauseButtonClicked before releasing the
MediaPlayer.

The AudioUIUpdater object is now declared in VoiceNotePlayer abstract class
instead of DefaultVoiceNotePlayer.

DefaultVoiceNotePlayer throws a better error message if a voice note is played
without first initializing the MediaPlayer.

In the rebindMonkeyItem method of MonkeyAdapter, if the monkeyholder is not
found, notifyItemChanged() is called. This seems to fix an issue with
MonkeyAudioHolders not updating.