Add conversations to UI Kit
Conversations have been added to the UI Kit in a similar way to messages. There
is a MonkeyConversation interface wich models the data; a
MonkeyConversationsAdapter that draws the conversations in the RecyclerView; a
MonkeyConversationsFragment to display the RecyclerView with the conversations
and a ConversationsActivity interface so that the fragment can talk to the
activity.

MainFragmentActivity now displays a short list of conversations in the sample
app. it has a new class FakeConversations to create the conversations and a
ConversationItem class to manipulate the conversations more easily.

Fixed a few issues with MonkeyTypesTest due to the recent changes to the  UI
Kit