Bind ConversationHolder to to MonkeyConversation
the onBindViewHolder of MonkeyConversationsAdapter has been implemented to draw
the conversations on the RecyclerView according the the conversation status and
other parameters.

The getItemViewType function of MonkeyConversationsAdapter has been implemented
to translate ConversationStatus enums to ConversationHolderTypes enums.

MonkeyConversationsAdapter has a new method insertConversations to add all
conversations.

ConversationHolder has a new constructor that receives the
ConversationHolderType to instantiate the view accordingly to the type.