Make fragments easier to use
New class MonkeyFragmentManager has methods to instantiate chat fragment and
conversation fragment in an easier way. This class adds the fragments to the
screen with animations that can be customized

MonkeyChatFragment now has a static constructor that adds the conversationId as
argument.

ChatActivity has new methods:

getInitialMessages() If the activity already has the first page of messages
loaded in memory, the fragment can retrieve them by calling this method before
setting the RecyclerView. If this method returns null, it calls onLoadMoreData()

retainMessages() the chat fragment will call this method on the onDestroy()
callback so that the activity can persist them after a configuration change.

MonkeyAdapter has new methods:

findMonkeyItemById() finds a MonkeyItem by id, starts by the end of the list.

clear() removes all MonkeyItem's from the adapter

messagesList is no longer private