Retain Conversations
ConversationsActivity interface has a new method retainConversations() to be
called in the onDestroy() callback so that the activity can retain the
conversations list on configuration changes.
Additionally, the setChatFragment method of MonkeyFragmentManager now returns a
list conversations. This is the list of displayed items. The activity must hold
on to it because when the conversations fragment comes back, it will have no
conversations.

Fixed an issue where MonkeyAdapter requested to upload files that did not exist.
In that case it should request to download instead.