Fix issues with adapter lists
conversationsList in MonkeyConversationsAdapter and messagesList in
MonkeyAdapter are private once again. MonkeyConversationsAdapter has a new
method takeAllConversations that removes the EndItem before returning the list.
MonkeyAdapter has new method takeAllMessages that works in a similar fashion.
This fixes a bug with loading views appearing twice.

New method in MonkeyConversationsAdapter: removeConversation()