Order messages before inserting
All methods that insert messages to MonkeyAdapter first calculate the correct
position of the new message so that the list is always ordered.
OrderedMessagesTest asserts that the list is always ordered

Dont try to show images for downloading incoming messages

Fix crash on rotating screen twice. Catch unintialized exception on
MonkeyConversationsFragment