Delete conversations on long click
After doing long click on a conversation in the RecyclerView, the conversation
is removed from the view and a snackbar is shown. Clicking on the undo button of
the snackbar brings back the conversation.
To include the snackbar, a new support design dependency has been added
Group conversations need an extra step. the must leave the group. When the
snackbar is removed from the screen, the new ConversationsActivity callback is
fired: onGroupLeft so that the activity can request the server to remove the
user from group

FakeConversations has been fixed to always have conversations with different
timestamps.

AttachmentButton Dialog has been moved to a new package Dialog so that other
classes can use it in the future

MonkeyConversationsAdapter now has a reference to the RecyclerView. This is
needed to update the UI on deletions.