Handle storage unavailable error
Photos are now stored in a special folder of Internal Storage. The default name
is "MonkeyKit Sent Photos"

PhotoEditTask has been modified to use a OnTaskFinished class as listener
instead of a Runnable. This makes it possible to notify the activity of any
exceptions raised during the editing and saving of the file. If the task can't
write the file, then it notifies the user and does not exit with RESULT_OK.

Even if PhotoEditTask manages to write the temporary file, there is no guarantee
that the bitmap will be written into the output file. For that casem InputListener
has been modified to include a new method: onNewItemFileError, to
notify the user that there was an error writing the file that he/she wanted to
send, most likely due to unavailable free space.