Handle unavailable space error on Photo Editor
CameraHandler now uses a tempFile and an outputFile instead of using a single
file reference for both things.
saveBitmapToFile no longer tries to catch an exception, instead it uses the
return value of the compress function to notify the AsyncTask whether the file
could be successfully saved or not.
new class OutputFile creates files in InternalStorage. This was moved from
CameraHandler so that audio can also use this.
PhotoEditorActivity now toggles progressBar visibilty with the processing flag.
Progress is also shown when processing bitmap before crop. Since the result of
PhotoEditTask is now a boolean, the value is used to decide whether to go on to
the next activity or show a toast informing the user that the file could not be
saved.