Reveal dialog after long click MonkeyHolder
MonkeyAdapter has new properties sentMessageOptions and receivedMessageOptions.
Each is a lost of OnMessageOptionClicked which hold callbacks to execute after
the user chooses an option.
When a monkeyHolder is long clicked a new Dialog is shown with the correct
options depending on the message type. the code to execute is implemented in
DefaultNameOptions.

Removed the checked image from all layouts.