Fix screen rotation issues
Camera handler's temporary files now always have the same filename. OutputFile
was modified to not use a timestamp to name temp files.
MonkeyConversationsFragment's conversationsAdapter is now nullable. the adapter
is set as null when the screen rotates while the conversations fragment is in
background.
When InputListener is set to MonkeyChatFragment, the new  reference is correctly
propagated all the way to MediaInputView's CameraHandler. the CameraHandler's
InputListener is also set in the initalize function instead of waiting for the
click event to do it. This allows CameraHandler to emit input events after
screen is rotated.
MonkeyAdapter's smoothlyAddNewItem now scrolls down on every message sent by the
user.