Center progressBar if RecyclerView is empty
When displaying progressBars while loading more items in a RecyclerView
with conversations or messages, the ProgressBar now takes the whole
height of the RecyclerView if the View is empty. This makes it look as
if the ProgressBar was centered.
This effect is achieved by manipulating the LayoutParams of the EndItem
view. during the onBind callback of the adapter.

To use this same logic for messages and conversations Utils has a new
function adjustHeight() changes the height of a view and can be used in
MonkeyAdapter and MonkeyConversationsAdapter.