Interface Player using VoiceNotePlayerBinder
Instead of passing a VoiceNotePlayer object to MonkeyAdapter to play
voice notes. A VoiceNotePlayerBinder is passed. the Binder object is an
interface to the service which should be the only source of truth for
the voice note player. PlaybackService is the only class that handles
VoiceNotePlayer objects, the rest have to use VoiceNotePlayerBinder
after binding to the service.