display file size on error views
When a file download/upload fails show the file size along with the
retry button. setErrorInUpload() and setErrorInDownload() now receive an
extra parameter, the size of the file.

xml files for transferring photos and voice notes have been modified to
easily access the TextView that will display the file size.