New class: ConversationsList
ConversationsList is meant to be the list that holds all the
conversations for the activity all the time without depending on a
MonkeyConversationsFragment instance. this object can temporarily hold a
referemce to a ConversationListUI object to update UI when the
conversation list changes.

MonkeyAdapter and MonkeyConversationsAdapter now use Snackbar.Callback
objects to execute code when the undo snackbar disappears. This is
possible now that we use taget API 23.

Fixed a bug when removing messages from MonkeyAdapter.

Updated SortedConversationsTest.