Refactor MonkeyFragmentManager
MonkeyFragmentManager has been modified to be easier to use.
MonkeyToolbar now deals only with the views of the toolbar.
MonkeyFragmentManager uses a stack to keep track of the fragments and
compares that to supportFragmentManager's backStack when the backstack
changes.

user must now call restoreToolbar() to draw the appropiate toolbar when
the activity is being restored.

MonkyChatFragment now has a Builder constructor for easier use.

UI Sample must now keep state in a headless fragment