Refactoring some behaviors for shippify Integration
InputViews now call ChatActivity's startMonkeyActivityForResult instead
of starting the activity directly so that host activity can execute
arbitrary code before switching activities.

The rules for scrolling in smoothlyAddNewItems are now more lenient.

new ConvList method: insertOrUpdateConversation
This Method tries to insert a conversation. If it already exists, it
replaces the current one

change notification text from conecting to syncing
In MonkeyStatusBar's showStatusNotification if status is not connected,
then just change the text in the notification. This fixes a bug that
notification would not show "syncing".

Deleting space between backbutton and headerview